pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "rag-system"
version = "1.0.0"
description = "Production-grade Retrieval-Augmented Generation system"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
keywords = ["rag", "llm", "retrieval", "vector-search", "nlp", "ai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "openai>=1.30.0",
    "tiktoken>=0.7.0",
    "sentence-transformers>=3.0.0",
    "faiss-cpu>=1.8.0",
    "chromadb>=0.5.0",
    "rank-bm25>=0.2.2",
    "pypdf>=4.2.0",
    "beautifulsoup4>=4.12.0",
    "python-docx>=1.1.0",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.7.0",
    "python-dotenv>=1.0.1",
    "tenacity>=8.3.0",
    "numpy>=1.26.0",
    "structlog>=24.1.0",
    "diskcache>=5.6.3",
    "tqdm>=4.66.0",
    "requests>=2.31.0",
    "lxml>=5.2.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "pytest-mock>=3.14",
    "ruff>=0.4.0",
    "black>=24.0",
    "mypy>=1.10",
    "types-requests",
]
eval = [
    "ragas>=0.1.9",
    "datasets>=2.19.0",
]
monitoring = [
    "prometheus-client>=0.20.0",
    "langfuse>=2.28.0",
]
cloud = [
    "pinecone-client>=3.2.0",
    "cohere>=5.0",
    "anthropic>=0.26.0",
]
all = [
    "rag-system[dev,eval,monitoring,cloud]",
]

[project.scripts]
rag-api = "api.main:main"

[project.urls]
Homepage       = "https://github.com/your-org/rag-system"
Documentation  = "https://github.com/your-org/rag-system/blob/main/ARCHITECTURE.md"
Repository     = "https://github.com/your-org/rag-system"
"Bug Tracker"  = "https://github.com/your-org/rag-system/issues"
Changelog      = "https://github.com/your-org/rag-system/blob/main/CHANGELOG.md"

# ── Tool configurations ───────────────────────────────────────────────────────
[tool.ruff]
line-length    = 99
target-version = "py310"
extend-select  = ["I", "N", "UP", "B", "SIM", "TCH"]
ignore         = [
    "B008",   # Do not perform function calls in default arguments
    "N806",   # Variable in function should be lowercase (conflicts with np, pd)
    "SIM108", # Use ternary operator instead of if-else
]
exclude = [
    ".git", "__pycache__", "venv", ".venv", "build", "dist", "notebooks",
]

[tool.ruff.per-file-ignores]
"tests/*"    = ["S101"]   # Allow assert in tests
"config.py"  = ["N815"]   # Allow mixed-case names in config

[tool.black]
line-length    = 99
target-version = ["py310", "py311", "py312"]
exclude        = '''
/(
    \.git
  | venv
  | \.venv
  | build
  | dist
  | notebooks
)/
'''

[tool.mypy]
python_version         = "3.11"
warn_return_any        = true
warn_unused_configs    = true
ignore_missing_imports = true
exclude                = ["venv/", "notebooks/", "tests/"]

[tool.pytest.ini_options]
testpaths        = ["tests"]
asyncio_mode     = "auto"
python_files     = ["test_*.py"]
python_classes   = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests that call external APIs (deselect with '-m not slow')",
    "integration: marks integration tests requiring a running server",
    "gpu: marks tests requiring GPU",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning:sentence_transformers",
]

[tool.coverage.run]
source   = ["."]
omit     = ["tests/*", "notebooks/*", "venv/*", "setup.py"]
branch   = true

[tool.coverage.report]
show_missing   = true
skip_covered   = false
fail_under     = 75
exclude_lines  = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
